+++ packages/core/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @rn-tools/core
 
+## 3.0.3
+
+### Patch Changes
+
+- Added render nodes to sheets
+
+## 3.0.2
+
+### Patch Changes
+
+- 723c379: Incorporate sheets into navigation and cleanup APIs
+
 ## 3.0.1
 
 ### Patch Changes